Remove DataView base class
